Fixed fonts formatting and querystring parsing
Fixed an issue where the default formatting was for Century Gothic font, and the backup font (Verdana) caused elements to overlap. Set the default font to Verdana to ease formatting pain. Fixed QueryString functionality such that the UI can now read query strings it is supplied and pass them on to the API. Allows URL sharing once functionality is added to modify URLs to reflect state.